requirements.txt
# FastAPI and Server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0
python-multipart>=0.0.9
supabase
python-dotenv
Jinja2
requests
razorpay  
email-validator
passlib
bcrypt==3.2.0  
google-auth
pandas>=2.2.3
numpy>=1.26.4
openpyxl>=3.1.2
matplotlib>=3.9.0
python-jose
scikit-learn
statsmodels